--- crates/fiber-lib/src/invoice/errors.rs
@@ -1,1 +1,0 @@
-// Re-exports removed. Use fiber_types::{InvoiceError, VerificationError} directly.
