+++ crates/fiber-lib/src/invoice/mod.rs
@@ -1,7 +1,5 @@
-mod errors;
 mod invoice_impl;
 mod store;
-mod utils;
 
 #[cfg(test)]
 mod tests;
